@@ -1,13 +1,12 @@
 source 'https://rubygems.org'
 
-
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '>=7.0.5.1'
+gem 'rails', '~> 8.0.1'
 # Use sqlite3 as the database for Active Record
-gem 'sqlite3'
+gem 'sqlite3', '>= 2.1'
 gem 'mysql2', '~> 0.5'
 # Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
+gem 'sassc-rails'
 # Use Uglifier as compressor for JavaScript assets
 gem 'uglifier', '>= 1.3.0'
 # See https://github.com/rails/execjs#readme for more supported runtimes
@@ -31,9 +30,10 @@ gem 'config'
 gem 'pdf-reader'
 
 # Rails server
-gem 'webrick', '~> 1.7'
+gem 'webrick', '>=1.8.2'
 
 gem 'rack-cors'
+gem 'lockbox', '=1.3.3'
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
