@@ -3,7 +3,9 @@ class UsersController < ApplicationController
 
   # return authenticated token upon signup
   def create
-    user = User.create!(user_params)
+    key = Lockbox.generate_key
+    my_user_params = user_params.to_h.merge(secret_key: key).merge(encryption_actived_flag: true)
+    user = User.create!(my_user_params)
     auth_token = AuthenticateUser.new(user.email, user.password).call
     response = { message: Message.account_created, auth_token: auth_token }
     json_response(response, :created)
